// src/library/models/src/lib/src/spells/phb/spell-names.ts
export enum Level1SpellNames {
  ABSORB_ELEMENTS = 'Absorb Elements',
  ALARM = 'Alarm',
  ANIMAL_FRIENDSHIP = 'Animal Friendship',
  ARMOR_OF_AGATHYS = 'Armor of Agathys',
  ARMS_OF_HADAR = 'Arms of Hadar',
  BANE = 'Bane',
  BLESS = 'Bless',
  BURNING_HANDS = 'Burning Hands',
  CEREMONY = 'Ceremony',
  CHARM_PERSON = 'Charm Person',
  CHROMATIC_ORB = 'Chromatic Orb',
  COLOR_SPRAY = 'Color Spray',
  COMMAND = 'Command',
  COMPELLED_DUEL = 'Compelled Duel',
  COMPREHEND_LANGUAGES = 'Comprehend Languages',
  CREATE_OR_DESTROY_WATER = 'Create or Destroy Water',
  CURE_WOUNDS = 'Cure Wounds',
  DETECT_EVIL_AND_GOOD = 'Detect Evil and Good',
  DETECT_MAGIC = 'Detect Magic',
  DETECT_POISON_AND_DISEASE = 'Detect Poison and Disease',
  DISGUISE_SELF = 'Disguise Self',
  DISSONANT_WHISPERS = 'Dissonant Whispers',
  DIVINE_FAVOR = 'Divine Favor',
  ENSNARING_STRIKE = 'Ensnaring Strike',
  ENTANGLE = ' Entangle',
  EXPEDITIOUS_RETREAT = 'Expeditious Retreat',
  FAERIE_FIRE = 'Faerie Fire',
  FALSE_LIFE = 'False Life',
  FEATHER_FALL = 'Feather Fall',
  FIND_FAMILIAR = 'Find Familiar',
  FOG_CLOUD = 'Fog Cloud',
  GOODBERRY = 'Goodberry',
  GREASE = 'Grease',
  GUIDING_BOLT = 'Guiding Bolt',
  HAIL_OF_THORNS = 'Hail of  Thorns',
  HEALING_WORD = 'Healing Word',
  HELLISH_REBUKE = 'Hellish Rebuke',
  HEROISM = 'Heroism',
  HEX = 'Hex',
  HUNTERS_MARK = 'Hunter\'s Mark',
  IDENTIFY = 'Identify',
  ILLUSORY_SCRIPT = 'Illusory Script',
  INFLICT_WOUNDS = 'Inflict Wounds',
  JUMP = 'Jump',
  LONGSTRIDER = 'Longstrider',
  MAGE_ARMOR = 'Mage Armor',
  MAGIC_MISSILE = 'Magic Missile',
  PROTECTION_FROM_EVIL_AND_GOOD = 'Protection from Evil and Good',
  PURIFY_FOOD_AND_DRINK = 'Purify Food and Drink',
  RAY_OF_SICKNESS = 'Ray of Sickness',
  SANCTUARY = 'Sanctuary',
  SEARING_SMITE = 'Searing Smite',
  SHIELD = 'Shield',
  SHIELD_OF_FAITH = 'Shield of Faith',
  SILENT_IMAGE = 'Silent Image',
  SLEEP = 'Sleep',
  SPEAK_WITH_ANIMALS = 'Speak with Animals',
  TASHAS_HIDEOUS_LAUGHTER = 'Tasha\'s Hideous Laughter',
  TENSERS_FLOATING_DISK = 'Tenser’s Floating Disk',
  THUNDERWAVE = 'Thunderwave',
  THUNDEROUS_SMITE = 'Thunderous Smite',
  UNSEEN_SERVANT = 'Unseen Servant',
  WITCH_BOLT = 'Witch Bolt',
  WRATHFUL_SMITE = 'Wrathful Smite'
}